+++ crates/ruma-client-api/src/lib.rs
@@ -12,6 +12,8 @@
 pub mod account;
 pub mod alias;
 pub mod appservice;
+#[cfg(feature = "unstable-msc3916")]
+pub mod authenticated_media;
 pub mod backup;
 pub mod config;
 pub mod context;