+++ crates/ruma-client-api/src/account/deactivate.rs
@@ -20,7 +20,7 @@ pub mod v3 {
     const METADATA: Metadata = metadata! {
         method: POST,
         rate_limited: true,
-        authentication: AccessToken,
+        authentication: AccessTokenOptional,
         history: {
             1.0 => "/_matrix/client/r0/account/deactivate",
             1.1 => "/_matrix/client/v3/account/deactivate",