+++ crates/ruma-client-api/src/authenticated_media/get_content_thumbnail.rs
@@ -2,10 +2,10 @@
 //!
 //! Get a thumbnail of content from the media store.
 
-pub mod unstable {
-    //! `/unstable/org.matrix.msc3916/` ([MSC])
+pub mod v1 {
+    //! `/v1/` ([spec])
     //!
-    //! [MSC]: https://github.com/matrix-org/matrix-spec-proposals/pull/3916
+    //! [spec]: https://spec.matrix.org/latest/client-server-api/#get_matrixclientv1mediathumbnailservernamemediaid
 
     use std::time::Duration;
 
@@ -24,6 +24,7 @@ pub mod unstable {
         authentication: AccessToken,
         history: {
             unstable => "/_matrix/client/unstable/org.matrix.msc3916/media/thumbnail/:server_name/:media_id",
+            1.11 => "/_matrix/client/v1/media/thumbnail/:server_name/:media_id",
         }
     };