+++ crates/ruma-client-api/CHANGELOG.md
@@ -18,6 +18,8 @@ Improvements:
 Bug fixes:
 
 - Rename `avatar` to `avatar_url` when (De)serializing
+- Make authentication with access token optional for the `change_password` and
+  `deactivate` endpoints.
 
 Bug fixes: